remove unisued imports, rename code file to match contents, etc.
