+++ asset_pile_viewer_app/lib/common/version_info.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 class VersionInfo {
   final int major;
   final int minor;
